# api/swagger.yaml
openapi: 3.0.0
info:
  title: Order Service API
  description: API for managing orders in the e-commerce platform
  version: 1.0.0
servers:
- url: https://localhost/v1
  description: Development server
paths:
  /orders:
    get:
      summary: List orders
      operationId: listOrders
      parameters:
      - name: userId
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: string
          format: uuid
      - name: status
        in: query
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/OrderStatus'
      - name: page
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 1
      - name: size
        in: query
        required: false
        style: form
        explode: true
        schema:
          type: integer
          default: 20
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "401":
          description: Unauthorized
      security:
      - apiKeyAuth: []
    post:
      summary: Create a new order
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
      security:
      - apiKeyAuth: []
  /orders/{orderId}:
    get:
      summary: Get an order by ID
      operationId: getOrder
      parameters:
      - name: orderId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "401":
          description: Unauthorized
        "404":
          description: Order not found
      security:
      - apiKeyAuth: []
    put:
      summary: Update an order
      operationId: updateOrder
      parameters:
      - name: orderId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
        required: true
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Order not found
      security:
      - apiKeyAuth: []
  /orders/{orderId}/cancel:
    post:
      summary: Cancel an order
      operationId: cancelOrder
      parameters:
      - name: orderId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Order not found
      security:
      - apiKeyAuth: []
components:
  schemas:
    OrderCreate:
      required:
      - billingAddress
      - items
      - shippingAddress
      - userId
      type: object
      properties:
        userId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        billingAddress:
          $ref: '#/components/schemas/Address'
    OrderUpdate:
      type: object
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        billingAddress:
          $ref: '#/components/schemas/Address'
    Order:
      type: object
      properties:
        orderId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        orderDate:
          type: string
          format: date-time
        totalAmount:
          type: number
          format: double
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        shippingAddress:
          $ref: '#/components/schemas/Address'
        billingAddress:
          $ref: '#/components/schemas/Address'
      example:
        totalAmount: 0.80082819046101150206595775671303272247314453125
        orderId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        shippingAddress:
          country: country
          city: city
          street: street
          postalCode: postalCode
          state: state
        billingAddress: null
        userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        orderDate: 2000-01-23T04:56:07.000+00:00
        items:
        - unitPrice: 1.46581298050294517310021547018550336360931396484375
          quantity: 1
          productId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - unitPrice: 1.46581298050294517310021547018550336360931396484375
          quantity: 1
          productId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        status: CREATED
    OrderItem:
      required:
      - productId
      - quantity
      - unitPrice
      type: object
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          minimum: 1
          type: integer
        unitPrice:
          type: number
          format: double
      example:
        unitPrice: 1.46581298050294517310021547018550336360931396484375
        quantity: 1
        productId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
    Address:
      required:
      - city
      - country
      - postalCode
      - state
      - street
      type: object
      properties:
        street:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
      example:
        country: country
        city: city
        street: street
        postalCode: postalCode
        state: state
    OrderStatus:
      type: string
      enum:
      - CREATED
      - PAID
      - SHIPPED
      - DELIVERED
      - CANCELLED
    inline_response_200:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        totalPages:
          type: integer
        totalElements:
          type: integer
      example:
        totalPages: 5
        content:
        - totalAmount: 0.80082819046101150206595775671303272247314453125
          orderId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          shippingAddress:
            country: country
            city: city
            street: street
            postalCode: postalCode
            state: state
          billingAddress: null
          userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          orderDate: 2000-01-23T04:56:07.000+00:00
          items:
          - unitPrice: 1.46581298050294517310021547018550336360931396484375
            quantity: 1
            productId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          - unitPrice: 1.46581298050294517310021547018550336360931396484375
            quantity: 1
            productId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: CREATED
        - totalAmount: 0.80082819046101150206595775671303272247314453125
          orderId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          shippingAddress:
            country: country
            city: city
            street: street
            postalCode: postalCode
            state: state
          billingAddress: null
          userId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          orderDate: 2000-01-23T04:56:07.000+00:00
          items:
          - unitPrice: 1.46581298050294517310021547018550336360931396484375
            quantity: 1
            productId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          - unitPrice: 1.46581298050294517310021547018550336360931396484375
            quantity: 1
            productId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
          status: CREATED
        totalElements: 5
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      name: X-API-KEY
      in: header